feat: Support for uiType === 'info'
'info' is for providing result data back to the frontend without pausing the state machine.
Processing continues to the next state. Useful for sending the results of a process that the user
might want to see, but for which no further action is needed by them (eg generating an export).